relocated component file locations for clarity
