+++ client/src/game/components/DescriptionBox.js
