+++ client/src/game/components/LowerBox.js
