+++ client/src/host/components/SelectionGrid.js
